chore: move cleanProposal.js
